requirements.txt
num2words
docx-mailmerge
PySimpleGUI
